fix the performance issue in create index with nchar/varchar/datetime
This problem is caused by the fmgr_security_definer keep searchCache for
function calls, and it accumulates to a considerable amount of time in
creating index since the function will be called for each record insert
into the btree.

Fix this problem by add the information into the cache, so it won't
searchCache every time when functions got called.

Task: BABEL-4097
Signed-off-by: Zhibai Song <szh@amazon.com>