Loss of decimal digits when converting from REAL to DECIMAL (#169) (#171)
Description
Initially we were using "%.*g" to write to buffer. This can use scientific notation as well. Now, FLT_DIG decides the number of digits after decimal. The value of FLT_DIG is 6. Which simply means the precision is 6. That is the reason why we are not observing any digits after decimal if number of digits before decimal is >= 6. Added a new macro for TSQL dialect named BABEL_MAX_PRECISION which will allow to accept precision of 38.

Issues Resolved
Task: BABEL-3066

3_X PR: #169

Signed-off-by: Shameem Ahmed [shmeeh@amazon.com]