Fix SELECT DISTINCT NULL result type
* Fix SELECT DISTINCT NULL result type

With distinct, Postgres coerces the type of unknown-typed targets so it
may find an equality function. It selects TEXT as the type to coerce to.

The problem in T-SQL is if we are unioning the result of SELECT DISTINCT
NULL with a different type. Some types, BIT for example, do not have
conversion functions between TEXT. Therefore, SELECT CAST(1 as BIT)
UNION SELECT DISTINCT NULL fails unexpectantly in Babelfish.

This change coerces NULL values to INT4 instead of TEXT when using
Babelfish, fixing the result type and the union behavior.

Signed-off-by: Walt Boettge <wboettge@amazon.com>
Task: BABEL-3392, BABEL-3348