Create DB Performance improvements (#234)
In Multi DB mode, as the number of databases increases, so does the time to create the next new DB.

This is because we create three internal roles for each new DB and internally when run the DB subcommands, multiple
calls to roles_is_member_of("sysadmin") is made. Now that output of this list contains all the three roles of every db
created. This is the major reason for the perfomance degradation of CREATE DB command.

We fix this in three different places.

getAvailDbid - this functions makes a call to nextval function, which by default checks for current user's permission and
makes a call to roles_is_member_of. Instead we could call the nextval_internal which is the same function but with the
additional option of check permissions flag which we will set to false. To double check we can just ensure that the
current user is "sysadmin" when getAvailDbid is called. (Currently we only call this when user is sysadmin)

Set temporary user when creating schema - when we create the dbo and guest schema for the new database, the
create schema function fetches all the roles that current role is member of (recursively) to check if if current role can
actually become the target schema owner role. To bypass this we can assume the newdb_dbo role when creating these
schemas. In this case all the roles that newdb_dbo is member of will be fetched, but this list is much smaller than
sysadmin.

Select best grantor - Select best grantor first fetches the roles_list that sysadmin is member of and then start checking
for permissions. But sysadmin is always the first to be checked. That is sysadmin is always top of the roles_list.
We can add a quick check to this. That is, first check if current role is sysadmin and can it give us all the permission
needed. If yes, simply return. Note** This does not change any behaviour since this will anyway be done in the first loop
after fetching roles_list. We are instead running the first loop before fetching the whole list.

Engine PR: #234
Extension PR: babelfish-for-postgresql/babelfish_extensions#1899
Extension PR: (cache sysadmin oid) babelfish-for-postgresql/babelfish_extensions#1942

Task: BABEL-4438
Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>