Support Instead of Trigger On Views
Currently Postgres does not support Instead of Trigger
On Views for DML queries.

On running such a query on VIEW, query rewriting substitutes
view with underlying base table.
Added a hook for skipping the substitution for DML
query on the VIEW.

This change will add support for Instead of Triggers
for DML queries when using Babelfish.

Task: BABEL-2170/BABEL-4532/BABEL-4514

Signed-off-by: Deepakshi Mittal <depakshi@amazon.com>